+++ src/general/RequestFactory.java
@@ -34,7 +34,8 @@ public class RequestFactory {
 		}
 		
 		Container_model t2 = Catalog.buyContainer();
-		float tmp = r.getToWAN()*(1-r.getImg_coeff())*r.getWs_as_coeff();
+		double tmp = r.getToWAN()*(1-r.getImg_coeff())*r.getWs_as_coeff();
+		
 		int M2 = (int)(tmp/Catalog.getContSpecs(t2)[2]);
 		ArrayList<Container> as = new ArrayList<Container> ();
 		for(int i=0;i < M2+1; i++) {
@@ -42,33 +43,37 @@ public class RequestFactory {
 		}
 		
 		Container_model t3 = Catalog.buyContainer();
-		float tmp2 = tmp*r.getAs_dbms_coeff();
+		double tmp2 = tmp*r.getAs_dbms_coeff();
+		
+
 		int M3 = (int)(tmp2/Catalog.getContSpecs(t3)[2]);
 		ArrayList<Container> dbms = new ArrayList<Container> ();
 		for(int i=0;i < M3+1; i++) {
 			dbms.add(new Container(t3,customer));
 		}
 		
-		 HashMap<C_Couple,Float> traffic = new HashMap<C_Couple,Float>();
+		 HashMap<C_Couple,Double> traffic = new HashMap<C_Couple,Double>();
 		 
 		 for(Container c: ws) {
-			 traffic.put(new C_Couple(Container.c_0,c), new Float(r.getFromWAN()/(M+1)));
-			 traffic.put(new C_Couple(c,Container.c_0), new Float(r.getToWAN()/(M+1)));
+			 traffic.put(new C_Couple(Container.c_0,c), new Double(r.getFromWAN()/(M+1)));
+			 traffic.put(new C_Couple(c,Container.c_0), new Double(r.getToWAN()/(M+1)));
 		 }
-	
+		 
+		// System.out.println("\n toWan = "+((float)r.getToWAN()/(M+1)));
 		for(Container c1 : ws) {
 			for(Container c2 : as) {
-				traffic.put(new C_Couple(c1,c2), new Float((r.getFromWAN()/(M+1))/(M2+1)) );
-				traffic.put(new C_Couple(c2,c1), new Float((tmp/(M+1))/(M2+1)));
+				traffic.put(new C_Couple(c1,c2), new Double((r.getFromWAN()/(M+1))/(M2+1)) );
+				traffic.put(new C_Couple(c2,c1), new Double((tmp/(M+1))/(M2+1)));
 			}
 		}
-	
+		// System.out.println("\n tmp = "+((float)(tmp/(M+1))/(M2+1)));
 		for(Container c1: as) {
 			for(Container c2: dbms) {
-				traffic.put(new C_Couple(c1,c2), new Float((r.getFromWAN()/(M2+1))/(M3+1)) );
-				traffic.put(new C_Couple(c2,c1), new Float((tmp2/(M2+1))/(M3+1)));
+				traffic.put(new C_Couple(c1,c2), new Double((r.getFromWAN()/(M2+1))/(M3+1)) );
+				traffic.put(new C_Couple(c2,c1), new Double((tmp2/(M2+1))/(M3+1)));
 			}
 		}
+		//System.out.println("\n tmp2 = "+((float)(tmp2/(M2+1))/(M3+1)));
 		return new Configuration(ws,as,dbms,traffic);
 	}
